% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixfile.R
\name{fixfile}
\alias{fixfile}
\title{Fix the output of 'Ahmes' 1.0}
\usage{
fixfile(x)
}
\arguments{
\item{x}{The output of 'Ahmes' 1.0. A .csv file}
}
\description{
Convert the output of 'Ahmes' 1.0 to a dataframe that can be handled by \code{'leafSTAR'} and \code{'R'} in general.
}
\details{
This function does not modify the original dataset. To have a permanent fixed version of the data, see Example. If using the new data within the package, be aware that the argument \code{Ahmes} should be set to \code{FALSE}.
}
\examples{
\donttest{data(olea)
 olea_fixed<-fixfile(olea) ## Store the fixed file as a new object.}
#fixfile()
}
\seealso{
\code{\link{Ahmes}}, \code{\link{fixhour}}
}
\author{
Agustina Ventre-Lespiaucq and Silvia Santamaria Bueno.
}
\keyword{'Ahmes'}
\keyword{fixers}
