% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixhour.R
\name{fixhour}
\alias{fixhour}
\title{Fix the hour format}
\usage{
fixhour(x, seconds = FALSE)
}
\arguments{
\item{x}{A character vector with the hour expressed as sexagesimal characters HH:MM or HH:MM:SS.}

\item{seconds}{Logical. Is the hour expressed as HH:MM:SS (\code{seconds=TRUE}) or as HH:MM (\code{seconds=FALSE})? Defaults to \code{FALSE}.}
}
\description{
Convert the hour from sexagesimal character to decimal numeric format.
}
\examples{
datahours<-c("00:30","10:00","18:20","20:55")
 datahours_fixed<-fixhour(datahours,seconds=FALSE)
#fixhour()
}
\seealso{
\code{\link{fixfile}}
}
\author{
Silvia Santamaria Bueno.
}
\keyword{fixers}
