% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropical.R
\docType{data}
\name{tropical}
\alias{tropical}
\title{Leaf angles of tropical canopy trees measured with traditional instrumentation}
\format{This data frame contains the following columns:

		\strong{species} A code for each species (letters) and individuals (numbers) 

		\strong{environment}  Factor with three levels of exposure to solar radiation.

		\strong{pitch} Pitch angle in degrees.

		\strong{roll} Roll angle in degrees.}
\usage{
tropical
}
\description{
The tropical data frame has 41 rows and 4 columns
}
\details{
These data are a subset of leaf angle data in Posada et al. (2009) measured with a compass and a protractor. Contains measurements of two leaf position angles of 41 leaves in two individuals of (\emph{Castilla elastica}) and one of (\emph{Ficus insipida}), each leaf in a row. The format of the data is an example of a field-data table handled by \code{'leafSTAR'}.
}
\references{
\code{tropical} \strong{Posada, J.M., Lechowicz, M.J., Kitajima, K. (2009)} Optimal photosynthetic use of light by tropical tree crowns achieved by adjustment of individual leaf angles and nitrogen content. Annals of Botany, 103: 795-805. doi:10.1093/aob/mcn265.
}
\keyword{datasets}
